Diplom/requirements.txt
selenium==4.33.0.
pytest==8.3.5.
allure-pytest==22.14.0
requests==2.31.
beautifulsoup4==4.12.2